# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tutor/ask:
    post:
      operationId: askTutor
      tags: [tutor]
      summary: Ask the AI tutor a question (text reply with optional spoken audio URL)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorAskInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorReply"
...
components:
  schemas:
...
    TutorAskInput:
      type: object
      required: [message]
      properties:
        sessionId: { type: ["integer", "null"] }
        message: { type: string }
        selectedLectureText: { type: ["string", "null"] }
        lectureId: { type: ["integer", "null"] }
        problemId: { type: ["integer", "null"] }
    TutorReply:
      type: object
      required: [text]
      properties:
        text: { type: string }
        audioUrl: { type: ["string", "null"] }
...